double 64

+++ src/lib/include/libKriging/Random.hpp
@@ -18,7 +18,7 @@ class Random {
 
   static void init();
 
-  LIBKRIGING_EXPORT static std::function<float()> randu;
+  LIBKRIGING_EXPORT static std::function<double()> randu;
 
   LIBKRIGING_EXPORT static std::function<arma::vec(const int)> randu_vec;
